+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change history for ui-developer
 
+## (IN PROGRESS)
+
+* Add Permissions Inspector page. Fixes UID-97.
+
 ## [6.0.0](https://github.com/folio-org/ui-developer/tree/v6.0.0) (2021-10-05)
 [Full Changelog](https://github.com/folio-org/ui-developer/compare/v5.2.1...v6.0.0